edit tables

+++ Cinema.SQL/dbo/Tables/Movies.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Movies] (
-    [Id]          BIGINT          IDENTITY (1, 1) NOT NULL,
+    [Id]          INT             IDENTITY (1, 1) NOT NULL,
     [Title]       NVARCHAR (150)  NOT NULL,
     [Description] NVARCHAR (1000) NOT NULL,
     [Poster]      VARBINARY (MAX) NULL,
@@ -7,4 +7,6 @@
     [EndDate]     DATE            NOT NULL,
     CONSTRAINT [PK_Movies] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
+
+
